belady: start of implementation of llc replacement
Belady's algorithm is simple - if you know the future. Luckily, we
can know the future if we read the whole trace before starting the
simulation. We record pairs of {address,sequence number} in a kind
of simple database (map of vectors) so that they can be looked up
later. When we must select a victim, we can look up all of the
existing addresses in the database to see which one will be used
farthest in the future.

Signed-off-by: Joel Nider <joel@nider.org>